Added ES 6.8 to RFS EndToEnd tests; refactored those tests
