+++ DocumentsFromSnapshotMigration/src/test/java/com/rfs/ParallelDocumentMigrationsTest.java
@@ -1,29 +1,19 @@
 package com.rfs;
 
 import java.nio.file.Files;
-import java.nio.file.Path;
-import java.time.Clock;
-import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.Random;
-import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Executors;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Function;
-import java.util.function.UnaryOperator;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import io.opentelemetry.sdk.metrics.data.LongPointData;
 import io.opentelemetry.sdk.metrics.data.MetricData;
-import org.apache.lucene.document.Document;
-import org.hamcrest.MatcherAssert;
-import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -34,36 +24,15 @@ import org.opensearch.migrations.metadata.tracing.MetadataMigrationTestContext;
 import org.opensearch.migrations.reindexer.tracing.DocumentMigrationTestContext;
 import org.opensearch.migrations.snapshot.creation.tracing.SnapshotTestContext;
 import org.opensearch.migrations.workcoordination.tracing.WorkCoordinationTestContext;
-import org.opensearch.testcontainers.OpensearchContainer;
 
-import com.rfs.cms.CoordinateWorkHttpClient;
-import com.rfs.cms.LeaseExpireTrigger;
-import com.rfs.cms.OpenSearchWorkCoordinator;
-import com.rfs.common.ClusterVersion;
-import com.rfs.common.DefaultSourceRepoAccessor;
-import com.rfs.common.DocumentReindexer;
 import com.rfs.common.FileSystemRepo;
 import com.rfs.common.FileSystemSnapshotCreator;
-import com.rfs.common.LuceneDocumentsReader;
 import com.rfs.common.OpenSearchClient;
-import com.rfs.common.RestClient;
-import com.rfs.common.SnapshotRepo;
-import com.rfs.common.SnapshotShardUnpacker;
-import com.rfs.common.SourceRepo;
-import com.rfs.common.SourceResourceProvider;
-import com.rfs.common.SourceResourceProviderFactory;
 import com.rfs.common.http.ConnectionContextTestParams;
 import com.rfs.framework.PreloadedSearchClusterContainer;
 import com.rfs.framework.SearchClusterContainer;
-import com.rfs.http.SearchClusterRequests;
-import com.rfs.models.IndexMetadata;
-import com.rfs.models.ShardMetadata;
-import com.rfs.worker.DocumentsRunner;
-import lombok.AllArgsConstructor;
 import lombok.Lombok;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import reactor.core.publisher.Flux;
 
 @Tag("longTest")
 @Slf4j
@@ -81,7 +50,6 @@ public class ParallelDocumentMigrationsTest extends SourceTestBase {
             .map(name -> makeParamsForBase(name))
             .collect(Collectors.toList());
         var targetImageNames = TARGET_IMAGES.stream()
-            .map(SearchClusterContainer.Version::getImageName)
             .collect(Collectors.toList());
         var numWorkersList = List.of(1, 3, 40);
         return sourceImageArgs.stream()
@@ -106,7 +74,7 @@ public class ParallelDocumentMigrationsTest extends SourceTestBase {
     @MethodSource("makeDocumentMigrationArgs")
     public void testDocumentMigration(
         int numWorkers,
-        String targetImageName,
+        SearchClusterContainer.Version targetVersion,
         SearchClusterContainer.Version baseSourceImageVersion,
         String generatorImage,
         String[] generatorArgs
@@ -125,7 +93,7 @@ public class ParallelDocumentMigrationsTest extends SourceTestBase {
                 generatorImage,
                 generatorArgs
             );
-            OpensearchContainer<?> osTargetContainer = new OpensearchContainer<>(targetImageName)
+            SearchClusterContainer osTargetContainer = new SearchClusterContainer(targetVersion);
         ) {
             CompletableFuture.allOf(CompletableFuture.supplyAsync(() -> {
                 esSourceContainer.start();
@@ -159,7 +127,7 @@ public class ParallelDocumentMigrationsTest extends SourceTestBase {
                     .build()
                     .toConnectionContext());
                 var sourceRepo = new FileSystemRepo(tempDir);
-                migrateMetadata(sourceRepo, targetClient, SNAPSHOT_NAME, INDEX_ALLOWLIST, testMetadataMigrationContext);
+                migrateMetadata(sourceRepo, targetClient, SNAPSHOT_NAME, List.of(), List.of(), List.of(), INDEX_ALLOWLIST, testMetadataMigrationContext, baseSourceImageVersion.getSourceVersion());
 
                 var workerFutures = new ArrayList<CompletableFuture<Integer>>();
                 var runCounter = new AtomicInteger();
@@ -172,7 +140,7 @@ public class ParallelDocumentMigrationsTest extends SourceTestBase {
                                 sourceRepo,
                                 SNAPSHOT_NAME,
                                 INDEX_ALLOWLIST,
-                                osTargetContainer.getHttpHostAddress(),
+                                osTargetContainer.getUrl(),
                                 runCounter,
                                 clockJitter,
                                 testDocMigrationContext,
@@ -264,170 +232,6 @@ public class ParallelDocumentMigrationsTest extends SourceTestBase {
         Assertions.assertEquals(numItemsAssigned, shardCount);
         long numCompleted = getMetricValueOrZero(workMetrics, "completeWorkCount");
         Assertions.assertEquals(numCompleted, shardCount + 1);
-    }
-
-    private void checkClusterMigrationOnFinished(
-        SearchClusterContainer esSourceContainer,
-        OpensearchContainer<?> osTargetContainer,
-        DocumentMigrationTestContext context
-    ) {
-        var targetClient = new RestClient(ConnectionContextTestParams.builder()
-            .host(osTargetContainer.getHttpHostAddress())
-            .build()
-            .toConnectionContext()
-        );
-        var sourceClient = new RestClient(ConnectionContextTestParams.builder()
-            .host(esSourceContainer.getUrl())
-            .build()
-            .toConnectionContext()
-        );
-
-        var requests = new SearchClusterRequests(context);
-        var sourceMap = requests.getMapOfIndexAndDocCount(sourceClient);
-        var refreshResponse = targetClient.get("_refresh", context.createUnboundRequestContext());
-        Assertions.assertEquals(200, refreshResponse.statusCode);
-        var targetMap = requests.getMapOfIndexAndDocCount(targetClient);
-
-        MatcherAssert.assertThat(targetMap, Matchers.equalTo(sourceMap));
-    }
-
-    @AllArgsConstructor
-    private static class ExpectedMigrationWorkTerminationException extends RuntimeException {
-        public final RfsMigrateDocuments.NoWorkLeftException exception;
-        public final int numRuns;
-    }
-
-    private int migrateDocumentsSequentially(
-        FileSystemRepo sourceRepo,
-        String snapshotName,
-        List<String> indexAllowlist,
-        String targetAddress,
-        AtomicInteger runCounter,
-        Random clockJitter,
-        DocumentMigrationTestContext testContext,
-        ClusterVersion parserVersion
-    ) {
-        for (int runNumber = 1;; ++runNumber) {
-            try {
-                var workResult = migrateDocumentsWithOneWorker(
-                    sourceRepo,
-                    snapshotName,
-                    indexAllowlist,
-                    targetAddress,
-                    clockJitter,
-                    testContext,
-                    parserVersion
-                );
-                if (workResult == DocumentsRunner.CompletionStatus.NOTHING_DONE) {
-                    return runNumber;
-                } else {
-                    runCounter.incrementAndGet();
-                }
-            } catch (RfsMigrateDocuments.NoWorkLeftException e) {
-                log.info(
-                    "No work at all was found.  "
-                        + "Presuming that work was complete and that all worker processes should terminate"
-                );
-                throw new ExpectedMigrationWorkTerminationException(e, runNumber);
-            } catch (Exception e) {
-                log.atError()
-                    .setCause(e)
-                    .setMessage(
-                        () -> "Caught an exception, "
-                            + "but just going to run again with this worker to simulate task/container recycling"
-                    )
-                    .log();
-            }
-        }
-    }
-
-    private static class FilteredLuceneDocumentsReader extends LuceneDocumentsReader {
-        private final UnaryOperator<Document> docTransformer;
-
-        public FilteredLuceneDocumentsReader(Path luceneFilesBasePath, boolean softDeletesPossible, String softDeletesField, UnaryOperator<Document> docTransformer) {
-            super(luceneFilesBasePath, softDeletesPossible, softDeletesField);
-            this.docTransformer = docTransformer;
-        }
-
-        @Override
-        public Flux<Document> readDocuments() {
-            return super.readDocuments().map(docTransformer::apply);
-        }
-    }
-
-    static class LeasePastError extends Error {}
-
-    @SneakyThrows
-    private DocumentsRunner.CompletionStatus migrateDocumentsWithOneWorker(
-        SourceRepo sourceRepo,
-        String snapshotName,
-        List<String> indexAllowlist,
-        String targetAddress,
-        Random clockJitter,
-        DocumentMigrationTestContext context,
-        ClusterVersion parserVersion
-    ) throws RfsMigrateDocuments.NoWorkLeftException {
-        var tempDir = Files.createTempDirectory("opensearchMigrationReindexFromSnapshot_test_lucene");
-        var shouldThrow = new AtomicBoolean();
-        try (var processManager = new LeaseExpireTrigger(workItemId -> {
-            log.atDebug().setMessage("Lease expired for " + workItemId + " making next document get throw").log();
-            shouldThrow.set(true);
-        })) {
-            UnaryOperator<Document> terminatingDocumentFilter = d -> {
-                if (shouldThrow.get()) {
-                    throw new LeasePastError();
-                }
-                return d;
-            };
-
-            SourceResourceProvider sourceResourceProvider = SourceResourceProviderFactory.getProvider(parserVersion);
-
-            DefaultSourceRepoAccessor repoAccessor = new DefaultSourceRepoAccessor(sourceRepo);
-            SnapshotShardUnpacker.Factory unpackerFactory = new SnapshotShardUnpacker.Factory(
-                repoAccessor,
-                tempDir,
-                sourceResourceProvider.getBufferSizeInBytes()
-            );
-
-            SnapshotRepo.Provider repoDataProvider = sourceResourceProvider.getSnapshotRepoProvider(sourceRepo);
-            IndexMetadata.Factory indexMetadataFactory = sourceResourceProvider.getIndexMetadataFactory(repoDataProvider);
-            ShardMetadata.Factory shardMetadataFactory = sourceResourceProvider.getShardMetadataFactory(repoDataProvider);
-            final int ms_window = 1000;
-            final var nextClockShift = (int) (clockJitter.nextDouble() * ms_window) - (ms_window / 2);
-            log.info("nextClockShift=" + nextClockShift);
-
-
-            Function<Path, LuceneDocumentsReader> readerFactory = path -> new FilteredLuceneDocumentsReader(path, sourceResourceProvider.getSoftDeletesPossible(),
-                sourceResourceProvider.getSoftDeletesFieldData(), terminatingDocumentFilter);
-
-            return RfsMigrateDocuments.run(
-                readerFactory,
-                new DocumentReindexer(new OpenSearchClient(ConnectionContextTestParams.builder()
-                    .host(targetAddress)
-                    .build()
-                    .toConnectionContext()), 1000, Long.MAX_VALUE, 1),
-                new OpenSearchWorkCoordinator(
-                    new CoordinateWorkHttpClient(ConnectionContextTestParams.builder()
-                        .host(targetAddress)
-                        .build()
-                        .toConnectionContext()),
-                    TOLERABLE_CLIENT_SERVER_CLOCK_DIFFERENCE_SECONDS,
-                    UUID.randomUUID().toString(),
-                    Clock.offset(Clock.systemUTC(), Duration.ofMillis(nextClockShift))
-                ),
-                Duration.ofMinutes(10),
-                processManager,
-                indexMetadataFactory,
-                snapshotName,
-                indexAllowlist,
-                shardMetadataFactory,
-                unpackerFactory,
-                MAX_SHARD_SIZE_BYTES,
-                context
-            );
-        } finally {
-            deleteTree(tempDir);
-        }
     }
 
 }
