+++ DocumentsFromSnapshotMigration/src/test/java/com/rfs/ProcessLifecycleTest.java
@@ -20,6 +20,7 @@ import org.opensearch.migrations.metadata.tracing.MetadataMigrationTestContext;
 import org.opensearch.migrations.snapshot.creation.tracing.SnapshotTestContext;
 import org.opensearch.testcontainers.OpensearchContainer;
 
+import com.rfs.common.ClusterVersion;
 import com.rfs.common.FileSystemRepo;
 import com.rfs.common.FileSystemSnapshotCreator;
 import com.rfs.common.OpenSearchClient;
@@ -121,7 +122,7 @@ public class ProcessLifecycleTest extends SourceTestBase {
             );
             esSourceContainer.copySnapshotData(tempDirSnapshot.toString());
 
-            migrateMetadata(osTargetContainer, tempDirSnapshot, testMetadataMigrationContext);
+            migrateMetadata(osTargetContainer, tempDirSnapshot, testMetadataMigrationContext, baseSourceImageVersion.getSourceVersion());
 
             int actualExitCode = runProcessAgainstToxicTarget(tempDirSnapshot, tempDirLucene, proxyContainer, failHow);
             log.atInfo().setMessage("Process exited with code: " + actualExitCode).log();
@@ -141,7 +142,8 @@ public class ProcessLifecycleTest extends SourceTestBase {
     private static void migrateMetadata(
         OpensearchContainer targetContainer,
         Path tempDirSnapshot,
-        MetadataMigrationTestContext testMetadataMigrationContext
+        MetadataMigrationTestContext testMetadataMigrationContext,
+        ClusterVersion sourceVersion
     ) {
         String targetAddress = "http://"
             + targetContainer.getHost()
@@ -152,7 +154,7 @@ public class ProcessLifecycleTest extends SourceTestBase {
             .build()
             .toConnectionContext());
         var sourceRepo = new FileSystemRepo(tempDirSnapshot);
-        migrateMetadata(sourceRepo, targetClient, SNAPSHOT_NAME, INDEX_ALLOWLIST, testMetadataMigrationContext);
+        migrateMetadata(sourceRepo, targetClient, SNAPSHOT_NAME, List.of(), List.of(), List.of(), INDEX_ALLOWLIST, testMetadataMigrationContext, sourceVersion);
     }
 
     private static int runProcessAgainstToxicTarget(
